+++ projects/core/orchestration/utils/k8s.py
@@ -1,0 +1,19 @@
+"""
+Kubernetes utilities for orchestration tasks
+"""
+
+from projects.core.dsl.utils.k8s import oc, resource_exists
+
+
+def ensure_namespace(namespace: str, *, labels: dict[str, str] | None = None) -> None:
+    """Ensure a namespace exists, creating it if necessary.
+
+    Args:
+        namespace: Namespace name
+        labels: Labels to apply to the namespace
+    """
+    if not resource_exists("namespace", namespace):
+        oc("create", "namespace", namespace)
+
+    if labels:
+        oc("label", "namespace", namespace, "--overwrite", *[f"{k}={v}" for k, v in labels.items()])
