--- projects/llm_d/orchestration/manifests/datasciencecluster.yaml
@@ -1,22 +1,0 @@
-apiVersion: datasciencecluster.opendatahub.io/v1
-kind: DataScienceCluster
-metadata:
-  name: default-dsc
-  namespace: redhat-ods-applications
-spec:
-  components:
-    codeflare:
-      managementState: Removed
-    dashboard:
-      managementState: Removed
-    datasciencepipelines:
-      managementState: Removed
-    kserve:
-      managementState: Managed
-      rawDeploymentServiceConfig: Headless
-    modelmeshserving:
-      managementState: Removed
-    ray:
-      managementState: Removed
-    workbenches:
-      managementState: Removed
