--- projects/llm_d/orchestration/manifests/nfd-nodefeaturediscovery.yaml
@@ -1,6 +1,0 @@
-apiVersion: nfd.openshift.io/v1
-kind: NodeFeatureDiscovery
-metadata:
-  name: nfd-instance
-  namespace: openshift-nfd
-spec: {}
